MRP: type change, msgbuf_len should also be size_t (no functional change)

+++ daemons/mrpd/mmrp.c
@@ -1269,7 +1269,7 @@ mmrp_emit_macvectors(unsigned char *msgbuf, unsigned char *msgbuf_eof,
 int mmrp_txpdu(void)
 {
 	unsigned char *msgbuf, *msgbuf_wrptr;
-	int msgbuf_len;
+	size_t msgbuf_len;
 	size_t bytes = 0;
 	eth_hdr_t *eth;
 	mrpdu_t *mrpdu;
@@ -1379,7 +1379,7 @@ int mmrp_txpdu(void)
 	} else
 		goto out;
 
-	msgbuf_len = (int)(mrpdu_msg_ptr - msgbuf);
+	msgbuf_len = mrpdu_msg_ptr - msgbuf;
 
 	bytes = mrpd_send(mmrp_socket, msgbuf, msgbuf_len, 0);
 #if LOG_MMRP
