+++ daemons/mrpd/msrp.c
@@ -2913,7 +2913,7 @@ msrp_emit_listenvectors(unsigned char *msgbuf, unsigned char *msgbuf_eof,
 int msrp_txpdu(void)
 {
 	unsigned char *msgbuf, *msgbuf_wrptr;
-	int msgbuf_len;
+	size_t msgbuf_len;
 	size_t bytes = 0;
 	eth_hdr_t *eth;
 	mrpdu_t *mrpdu;
@@ -2990,7 +2990,7 @@ int msrp_txpdu(void)
 	} else
 		goto out;
 
-	msgbuf_len = (int)(mrpdu_msg_ptr - msgbuf);
+	msgbuf_len = mrpdu_msg_ptr - msgbuf;
 
 	bytes = mrpd_send(msrp_socket, msgbuf, msgbuf_len, 0);
 #if LOG_MSRP
